requirements.txt
asttokens==3.0.0
bitarray==3.0.0
bitstring==4.3.0
bleak==0.22.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
enum-compat==0.0.3
executing==2.2.0
fonttools==4.56.0
future==1.0.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyboard==0.13.5
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
muselsl==2.3.1
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pygatt==5.0.0
Pygments==2.19.1
pylsl==1.17.6
pynput==1.8.0
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.1
wcwidth==0.2.13
winrt-runtime==2.3.0
winrt-Windows.Devices.Bluetooth==2.3.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.3.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.3.0
winrt-Windows.Devices.Enumeration==2.3.0
winrt-Windows.Foundation==2.3.0
winrt-Windows.Foundation.Collections==2.3.0
winrt-Windows.Storage.Streams==2.3.0
